fastapi-mcp-server/pyproject.toml
[project]
name = "fastapi-mcp-server"
version = "0.1.0"
description = "FastAPI MCP Server for documentation indexing and retrieval"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.9.0",
    "elasticsearch>=8.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "markdown>=3.5.0",
    "beautifulsoup4>=4.12.0",
    "gitpython>=3.1.0",
    "aiofiles>=23.0.0",
    "aiohttp>=3.8.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
]
